% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{det_outcome}
\alias{det_outcome}
\title{Function to determine whether to sample from binary outcome or the real line}
\usage{
det_outcome(var, treat_effect = NULL, x = NULL)
}
\description{
Function to determine whether to sample from binary outcome or the real line
}
