% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{gen_errormat}
\alias{gen_errormat}
\title{Function to generate error matrix}
\usage{
gen_errormat(data, time.ac, spatial.ac)
}
\description{
This function generates a distribution of errors given a set of 
over-time or between-cases coefficients. Errors can be affected by 
either spatial or time autocorrelation.
}
