% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_parallel.R
\name{mclapply}
\alias{mclapply}
\title{Overwrite the serial version of mclapply on Windows only}
\usage{
mclapply(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE,
  affinity.list = NULL
)
}
\arguments{
\item{empty}{it takes nothing}
}
\value{
mclapply like list
}
\description{
Overwrite the serial version of mclapply on Windows only
}
