% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_parallel.R
\name{mclapply_socket}
\alias{mclapply_socket}
\title{Define a sockets version of mclapply}
\usage{
mclapply_socket(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = NULL, mc.cleanup = TRUE,
  mc.allow.recursive = TRUE)
}
\arguments{
\item{...}{What you pass to mclapply}
}
\value{
mclapply like list
}
\description{
An implementation of \code{\link[parallel]{mclapply}} using \code{parallel::parLapply}
}
\details{
Windows does not support forking. This makes it impossible to use mclapply on Windows to
farm out work to additional cores.

More words
}
