% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{tw_data}
\alias{tw_data}
\title{Generate One-way and Two-way Fixed Effects Panel Data}
\usage{
tw_data(
  N = 30,
  T = 30,
  case.int.mean = 0.5,
  case.int.sd = 0.5,
  cross.int.mean = -1,
  cross.int.sd = 0.5,
  cross.eff.mean = 0,
  did.eff.mean = 1,
  did.eff.sd = 0.25,
  wid.eff.mean = 0,
  wid.eff.sd = 0,
  cross.eff.sd = 0.5,
  case.eff.mean = 0.5,
  case.eff.sd = 0.5,
  noise.sd = 1,
  omm.x.case = 0,
  omm.x.cross = 0,
  omm.y.case = 0,
  omm.y.cross = 0,
  treat_effect = NULL,
  this_case = NULL,
  this_time = NULL,
  binary_outcome = FALSE,
  binary_x = FALSE,
  unbalance = FALSE,
  time.ac = 0,
  spatial.ac = 0,
  prior_true_vals = NULL
)
}
\arguments{
\item{N}{The number of observations for each case/unit.}

\item{T}{The number of time points per observation.}

\item{case.int.mean}{The mean of the case/unit intercepts/fixed effects}

\item{case.int.sd}{The SD of the case/unit intercepts/fixed effects}

\item{cross.int.mean}{The mean of the cross-sectional intercepts/fixed effects}

\item{cross.int.sd}{The SD of the cross-sectional intercepts/fixed effects}

\item{cross.eff.mean}{The mean of the cross-sectional effect of X on Y}

\item{did.eff.mean}{The mean of the difference-in-difference effect of X on Y}

\item{did.eff.sd}{The SD of the difference-in-difference effect of X on Y}

\item{wid.eff.mean}{The mean of the difference-in-cases effect of X on Y}

\item{wid.eff.sd}{The SD of the difference-in-cases effect of X on Y}

\item{cross.eff.sd}{The SD of the cross-sectional effect of X on Y}

\item{case.eff.mean}{The mean of the case (over-time) effect of X on Y}

\item{case.eff.sd}{The SD of the case (over-time) effect of X on Y}

\item{noise.sd}{The residual variance of the data}

\item{omm.x.case}{The value of an omitted variable correlated with X that varies across cases/units}

\item{omm.x.cross}{The value of an omitted variable correlated with X that varies cross-sectionally}

\item{omm.y.case}{The value of an omitted variable correlated with Y that varies across cases/units}

\item{omm.y.cross}{The value of an omitted variable correlated with Y that varies cross-sectionally}

\item{treat_effect}{A vector of length 1 or N*T that is equal to 1 for assignment to treatment and 0 for assignment to control. If this argument is not NULL, `tw_data` will generate data for y that takes x as fixed to these values (i.e., treatment is being assigned/manipulated).}

\item{binary_outcome}{Whether the Y (outcome) variable should be converted to 0/1. Note that this can lead to some measurement bias as Y is simulated as continuous.}

\item{binary_x}{Whether X should be converted to 0/1. Note that doing so may lead to some bias in estimation as X is simulated as continuous.}

\item{unbalance}{Whether to simulate varying numbers of observations by cases or time points.}

\item{time.ac}{A value between 0 and 1 giving the over-time autocorrelation in effect of X on Y}

\item{spatial.ac}{A value between 0 and 1 giving the cross-sectional (spatial)
autocorrelation in the effect of X on Y}

\item{prior_true_vals}{A fitted `tw_data` object with generated coefficients that can be used to keep vectors of intercepts/effects fixed over repeated sampling}
}
\value{
The function returns a named list where \code{object$data} is a \code{data.frame} and
\code{object$pars} are the original parametes used to generate the data. The value of generated coefficients is returned in a list as \code{object$fixed_params}. For repeated sampling, the object can be given to the `prior_true_vals` argument to allow for fixed population parameters for intercepts and vectors of effects.
}
\description{
This function will produce panel data where variation can exist in the cross-section, over time
or in both dimensions simultaneously. Furthermore, effect heterogeneity by case or cross section
is also allowed, along with interactive effects such as differences-in-differences.
}
\details{
The \code{tw_data} function is the workhorse of the \code{twowaysim} package. It accepts as
input the dimensions of the panel/TSCS data to be generated, and also parameters that
determine the extent of variance and heterogeneity in either the cross-sectional or
over-time effects in the data or the interaction thereof. The parameter \code{N} determines how many observations
exist for each case or unit in the panel, while \code{T} determines how many time points exist
per case or unit. To create a model with a homogenous (static) within-unit over-time (case) effect,
simply set \code{case.eff.mean} to a non-zero number and set \code{case.eff.sd} to zero. Similarly,
setting \code{cross.eff.mean} to a non-zero number and \code{cross.eff.sd} to zero will produce a
panel dataset with a cross-sectional effect of X on Y where the effect of X does not vary across
countries (no effect heterogeneity). Increasing \code{cross.eff.sd} and \code{case.eff.sd} will result in more
effect heterogeneity across countries and time points. If both \code{case.eff.mean} and \code{cross.eff.mean} are
non-zero, then Y will have both dimensions of variance. A 1-way fixed effects model with intercepts on
cases will return the \code{case.eff.mean} coefficient and a model with intercepts on time points will return
the \code{cross.eff.mean} estimate, whereas a 2-way model (intercepts on cases and time points) will return
a difficult-to-characterize weighted average.

To estimate a difference-in-differences (time interaction) effect, set \code{did.eff.mean} to a non-zero number and set \code{did.eff.sd} to zero if the DiD effects are supposed to be homogenous. Note that to estimate a standard "canonical" DiD setup, the time points \code{T} should be no more than 2. For more information about generating DiD specifications with `tw_data`, see the blog post \url{https://www.robertkubinec.com/post/did_dnd/index.html}.

We refer you to Kropko and Kubinec (2020) for more information
on the difference between these models:
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0231349}.

The parameters \code{case.int} and \code{cross.int} represent the values of the intercepts for the
cases or time points. Changing these parameters will increase or decrease the amount of unexplained
variance (random noise) in the dataset.

The additional parameters in the function allow the user to create unbalanced panels (varying numbers
of observations per case or time point if \code{unbalance=TRUE}), auto-correlation in the effects and
omitted variables. Autocorrelation can exist either in the over time dimension or the cross-sectional
dimension. To increase time autocorrelation, set \code{time.ac} to a value between 0 and 1 where values
closer to one signal higher autocorrelation. To increase spatial (cross-sectional) autocorrelation,
set \code{spatial.ac} to a value between 0 and 1.

Finally, to include omitted variables, set one of the \code{omm} parameters to a non-zero value.
Omitted variables that vary within cases (over time) can be included by setting an \code{omm} parameter
subscripted with \code{case} to a non-zero value, and the same is possible for variables that vary
in the cross-section \code{cross}. The analyst can also decide whether the omitted variable is correlated
with the independent variable of interest \code{x} or the dependent variable \code{y} by choosing the
subscript of \code{omm}.
}
\examples{

# case (over-time) effect with no effect heterogeneity

case1 <- tw_data(case.eff.mean=-1,case.eff.sd=0)

# case (over-time) effect with substantial effect heterogeneity across countries

case2 <- tw_data(case.eff.mean=-1,case.eff.sd=1)

# cross-section effect with no effect heterogeneity

cross1 <- tw_data(cross.eff.mean=-1,cross.eff.sd=0)

# cross-section effect with substantial effect heterogeneity across countries

cross2 <- tw_data(cross.eff.mean=-1,cross.eff.sd=1)

# panel data with a cross-sectional effect of 3 and a case (over-time) effect of -1

both_case_cross <- tw_data(cross.eff.mean=3,
                             case.eff.mean=-1)

}
\seealso{
\code{\link{tw_model}} for running linear models on the data and and
\code{\link{tw_sim}} function for running Monte Carlo simulations on panel data.
}
