% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{tw_plot}
\alias{tw_plot}
\title{Plot Results of \code{tw_sim} simulations}
\usage{
tw_plot(gensim, use_ci = T, ...)
}
\arguments{
\item{gensim}{A set of simulations produced by \code{\link{tw_sim}}}

\item{use_ci}{Whether to add confidence intervals around LOESS curves.
Turn off if the CIs are too big relative to plot.}

\item{...}{Currently ignored}
}
\description{
This function will generate plots comparing the performance of different estimators
at returning \code{\link{tw_sim}} Monte Carlo simulations.
}
