% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{tw_data}
\alias{tw_data}
\title{Generate One-way and Two-way Fixed Effects Panel Data}
\usage{
tw_data(N = 30, T = 30, case.int.mean = 0, case.int.sd = 1,
  cross.int.mean = 0, cross.int.sd = 1, cross.eff.mean = 3,
  cross.eff.sd = 0.5, case.eff.mean = -3, case.eff.sd = 0.5,
  noise.sd = 1, omm.x.case = 0, omm.x.cross = 0, omm.y.case = 0,
  omm.y.cross = 0, binary = FALSE, unbalance = FALSE, time.ac = 0,
  spatial.ac = 0)
}
\arguments{
\item{N}{The number of observations for each case/unit.}

\item{T}{The number of time points per observation.}

\item{case.int.mean}{The mean of the case/unit intercepts/fixed effects}

\item{case.int.sd}{The SD of the case/unit intercepts/fixed effects}

\item{cross.int.mean}{The mean of the cross-sectional intercepts/fixed effects}

\item{cross.int.sd}{The SD of the cross-sectional intercepts/fixed effects}

\item{cross.eff.mean}{The mean of the cross-sectional effect of X on Y}

\item{cross.eff.sd}{The SD of the cross-sectional effect of X on Y}

\item{case.eff.mean}{The mean of the case (over-time) effect of X on Y}

\item{case.eff.sd}{The SD of the case (over-time) effect of X on Y}

\item{noise.sd}{The residual variance of the data}

\item{omm.x.case}{The value of an omitted variable correlated with X that varies across cases/units}

\item{omm.x.cross}{The value of an omitted variable correlated with X that varies cross-sectionally}

\item{omm.y.case}{The value of an omitted variable correlated with Y that varies across cases/units}

\item{omm.y.cross}{The value of an omitted variable correlated with Y that varies cross-sectionally}

\item{binary}{Whether to generate a continuous outcome/DV or a binary outcome/DV}

\item{unbalance}{Whether to simulate varying numbers of observations by cases or time points.}

\item{time.ac}{A value between 0 and 1 giving the over-time autocorrelation in effect of X on Y}

\item{spatial.ac}{A value between 0 and 1 giving the cross-sectional (spatial) 
autocorrelation in the effect of X on Y}
}
\value{
The function returns a named list where \code{object$data} is a \code{data.frame} and 
\code{object$pars} are the original parametes used to generate the data.
}
\description{
This function will produce panel data where variation can exist in the cross-section, over time 
or in both dimensions simultaneously. Furthermore, effect heterogeneity by case or cross section
is also allowed.
}
\details{
The \code{tw_data} function is the workhorse of the \code{twowaysim} package. It accepts as 
input the dimensions of the panel/TSCS data to be generated, and also parameters that 
determine the extent of variance and heterogeneity in either the cross-sectional or 
over-time effects in the data. The parameter \code{N} determines how many observations
exist for each case or unit in the panel, while \code{T} determines how many time points exist 
per case or unit. To create a model with a within-unit over-time (case) effect,
simply set \code{case.eff.mean} to a non-zero number and set \code{case.eff.sd} to zero. Similarly, 
setting \code{cross.eff.mean} to a non-zero number and \code{cross.eff.sd} to zero will produce a 
panel dataset with a cross-sectional effect of X on Y where the effect of X does not vary across
countries (no effect heterogeneity). Increasing \code{cross.eff.sd} and \code{case.eff.sd} will result in more
effect heterogeneity across countries and time points. If both \code{case.eff.mean} and \code{cross.eff.mean} are
non-zero, then Y will have both dimensions of variance. A 1-way fixed effects model with intercepts on 
cases will return the \code{case.eff.mean} coefficient and a model with intercepts on time points will return
the \code{cross.eff.mean} estimate, whereas a 2-way model (intercepts on cases and time points) will return 
a difficult-to-characterize weighted average.

We refer you to Kropko and Kubinec (2018) for more information
on the difference between these models: 
\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3062619}.

The parameters \code{case.int} and \code{cross.int} represent the values of the intercepts for the
cases or time points. Changing these parameters will increase or decrease the amount of unexplained
variance (random noise) in the dataset.

The additional parameters in the function allow the user to create unbalanced panels (varying numbers
of observations per case or time point if \code{unbalance=TRUE}), auto-correlation in the effects and 
omitted variables. Autocorrelation can exist either in the over time dimension or the cross-sectional 
dimension. To increase time autocorrelation, set \code{time.ac} to a value between 0 and 1 where values
closer to one signal higher autocorrelation. To increase spatial (cross-sectional) autocorrelation, 
set \code{spatial.ac} to a value between 0 and 1.

Finally, to include omitted variables, set one of the \code{omm} parameters to a non-zero value. 
Omitted variables that vary within cases (over time) can be included by setting an \code{omm} parameter
subscripted with \code{case} to a non-zero value, and the same is possible for variables that vary
in the cross-section \code{cross}. The analyst can also decide whether the omitted variable is correlated
with the independent variable of interest \code{x} or the dependent variable \code{y} by choosing the
subscript of \code{omm}.
}
\examples{

# case (over-time) effect with no effect heterogeneity

case1 <- tw_data(case.eff.mean=-1,case.eff.sd=0)

# case (over-time) effect with substantial effect heterogeneity across countries

case2 <- tw_data(case.eff.mean=-1,case.eff.sd=1)

# cross-section effect with no effect heterogeneity

cross1 <- tw_data(cross.eff.mean=-1,cross.eff.sd=0)

# cross-section effect with substantial effect heterogeneity across countries

cross2 <- tw_data(cross.eff.mean=-1,cross.eff.sd=1)

# panel data with a cross-sectional effect of 3 and a case (over-time) effect of -1

both_case_cross <- tw_data(cross.eff.mean=3,
                             case.eff.mean=-1)

}
\seealso{
\code{\link{tw_model}} for running linear models on the data and and 
\code{\link{tw_sim}} function for running Monte Carlo simulations on panel data.
}
