% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsim2.R
\name{tw_plot}
\alias{tw_plot}
\title{Plot Results of \code{tw_sim} simulations}
\usage{
tw_plot(
  gensim,
  display_est = "coef",
  use_ci = T,
  xvar = NULL,
  yvar = NULL,
  facet_scales = "fixed",
  ...
)
}
\arguments{
\item{gensim}{A set of simulations produced by \code{\link{tw_sim}}}

\item{display_est}{Whether to display estimated coefficients of
independent variables from the linear model (\code{"coef"}) or the
standard error of these estimates (\code{"se"})}

\item{use_ci}{Whether to add confidence intervals around LOESS curves.
Turn off if the CIs are too big relative to plot.}

\item{xvar}{In case of multiple attributes varied in the simulation, which
attribute should be on the x axis? See \code{\link{tw_data}} for possible
attribute names. Should be an unquoted variable name like
\code{case.int.sd}.}

\item{yvar}{In case of multiple attributes varied in the simulation, which
attribute should be on the y axis? See \code{\link{tw_data}} for possible
attribute names. Should be an unquoted variable name like
\code{case.int.sd}.}

\item{facet_scales}{The value of \code{scales} option passed to the
\code{ggplot2} \code{facet_wrap} function to control the scales.
Changes to \code{'free_x'},\code{'free_y'} or \code{'free'} if the
scales obscure comparison}

\item{...}{Currently ignored}
}
\description{
This function will generate plots comparing the performance of different estimators
at returning \code{\link{tw_sim}} Monte Carlo simulations.
}
